+++ grafana_alloy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.11] - 2026-08-30
+
+- Added `environment_variables` option to set environment variables for the Alloy process, usable in the config with `sys.env()` [#18](https://github.com/wymangr/hassos-addons/issues/18)
+- Bump Grafana Alloy version to [1.19.2](https://github.com/grafana/alloy/releases/tag/v1.19.2)
+- Bump base image from `9.3.0` to `9.4.0`
+
 ## [0.0.10] - 2026-08-07
 
 - Publish pre-built images to GHCR so the add-on is pulled instead of built on the device